pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "drone-search"
version = "0.1.0"
description = "Multimodal person retrieval and triage over aerial drone video. JHU EN.601.466/666 final project (Spring 2026)."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Timothy Lin", email = "timothylinziqi@gmail.com" }]

dependencies = [
    "ultralytics>=8.2.0",
    "open-clip-torch>=2.24.0",
    "torch>=2.2.0",
    "faiss-cpu>=1.8.0",
    "scikit-learn>=1.4.0",
    "streamlit>=1.32.0",
    "pillow>=10.0.0",
    "opencv-python-headless>=4.9.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "whoosh>=2.7.4",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
drone-search = "drone_search.__main__:app"

[tool.hatch.build.targets.wheel]
packages = ["src/drone_search"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
